requirements.txt
google-genai
transformers
torch
langdetect
textblob
